task no7: correct bussiness logic and try to test overall input and output

+++ src/main/java/com/qinnnyul/pos/domain/ShoppingItem.java
@@ -3,9 +3,9 @@ package com.qinnnyul.pos.domain;
 public class ShoppingItem {
     private String name;
     private double price;
-    private int amount;
+    private double amount;
 
-    public ShoppingItem(String name, double price, int amount) {
+    public ShoppingItem(String name, double price, double amount) {
         this.name = name;
         this.price = price;
         this.amount = amount;
@@ -19,7 +19,7 @@ public class ShoppingItem {
         return price;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
